Initial System Generator Working
I made the generator for the initial system. It works as we wanted it to, but there are a few things I might want to add to it in order to make it a little nicer. You guys can use this to do any tests on your systems, and give me feedback if you see anything weird about it.